+++ ntex/src/http/h2/service.rs
@@ -6,7 +6,7 @@ use ntex_h2::{self as h2, frame::StreamId, server};
 use crate::http::body::{BodySize, MessageBody};
 use crate::http::config::{DispatcherConfig, ServiceConfig};
 use crate::http::error::{DispatchError, H2Error, ResponseError};
-use crate::http::header::{self, HeaderMap, HeaderValue};
+use crate::http::header::{self, HeaderMap, HeaderName, HeaderValue};
 use crate::http::message::{CurrentIo, ResponseHead};
 use crate::http::{DateService, Method, Request, Response, StatusCode, Uri, Version};
 use crate::io::{types, Filter, Io, IoBoxed, IoRef};
@@ -440,39 +440,46 @@ where
     }
 }
 
-fn prepare_response(timer: &DateService, head: &mut ResponseHead, size: &mut BodySize) {
-    let mut skip_len = size == &BodySize::Stream;
+#[allow(clippy::declare_interior_mutable_const)]
+const ZERO_CONTENT_LENGTH: HeaderValue = HeaderValue::from_static("0");
+#[allow(clippy::declare_interior_mutable_const)]
+const KEEP_ALIVE: HeaderName = HeaderName::from_static("keep-alive");
+#[allow(clippy::declare_interior_mutable_const)]
+const PROXY_CONNECTION: HeaderName = HeaderName::from_static("proxy-connection");
 
+fn prepare_response(timer: &DateService, head: &mut ResponseHead, size: &mut BodySize) {
     // Content length
     match head.status {
         StatusCode::NO_CONTENT | StatusCode::CONTINUE | StatusCode::PROCESSING => {
             *size = BodySize::None
         }
         StatusCode::SWITCHING_PROTOCOLS => {
-            skip_len = true;
             *size = BodySize::Stream;
-            head.headers.remove(header::CONTENT_LENGTH);
         }
         _ => (),
     }
-    let _ = match size {
-        BodySize::None | BodySize::Stream => (),
+    match size {
+        BodySize::None | BodySize::Stream => head.headers.remove(header::CONTENT_LENGTH),
         BodySize::Empty => head
             .headers
-            .insert(header::CONTENT_LENGTH, HeaderValue::from_static("0")),
+            .insert(header::CONTENT_LENGTH, ZERO_CONTENT_LENGTH),
         BodySize::Sized(len) => {
-            if !skip_len {
-                head.headers.insert(
-                    header::CONTENT_LENGTH,
-                    HeaderValue::try_from(format!("{}", len)).unwrap(),
-                );
-            }
+            head.headers.insert(
+                header::CONTENT_LENGTH,
+                HeaderValue::try_from(format!("{}", len)).unwrap(),
+            );
         }
     };
 
     // http2 specific1
     head.headers.remove(header::CONNECTION);
     head.headers.remove(header::TRANSFER_ENCODING);
+    head.headers.remove(header::UPGRADE);
+
+    // omit HTTP/1.x only headers according to:
+    // https://datatracker.ietf.org/doc/html/rfc7540#section-8.1.2.2
+    head.headers.remove(KEEP_ALIVE);
+    head.headers.remove(PROXY_CONNECTION);
 
     // set date header
     if !head.headers.contains_key(header::DATE) {